# ta-api-master/swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/flight/order-detail": {
      "post": {
        "operationId": "OrderDetailController_getOrderDetails",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "header",
            "description": "Please enter preferred language. (e.g., english,french)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          },
          {
            "name": "Club-Id",
            "in": "header",
            "description": "Please enter club id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Device-Information",
            "in": "header",
            "description": "Please enter device-information id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDetailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get order details",
        "tags": [
          "Flight"
        ]
      }
    },
    "/flight/book/initiate": {
      "post": {
        "operationId": "BookController_bookingInitiate",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Book a flight",
        "tags": [
          "Flight"
        ]
      }
    },
    "/flight/book/confirmation": {
      "post": {
        "operationId": "BookController_bookingConfirmation",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookConfirmationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Book a flight",
        "tags": [
          "Flight"
        ]
      }
    },
    "/flight/revalidate": {
      "post": {
        "operationId": "RevalidateController_revalidate",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevalidateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Revalidate a flight",
        "tags": [
          "Flight"
        ]
      }
    },
    "/flight/balance-check": {
      "post": {
        "operationId": "BalanceCheckController_balanceCheck",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BalanceCheckDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Check balance",
        "tags": [
          "Flight"
        ]
      }
    },
...
  },
  "info": {
    "title": "API SERVICE",
    "description": "## Travel API Documentation\n\nWelcome to the **API SERVICE** documentation. This guide provides comprehensive guidelines and detailed instructions on how to effectively integrate and utilize the Travel API.\n\n### Overview\nThis documentation includes various sections covering:\n- **Authentication Methods**: Learn how to authenticate using Bearer and Cookie authentication.\n- **Endpoint Descriptions**: Detailed information on each API endpoint.\n- **Usage Examples**: Practical examples to help you get started quickly.\n- **Best Practices**: Tips and recommendations for efficient implementation.\n\n### Purpose\nWhether you are a developer looking to incorporate travel-related functionalities or a business aiming to enhance your travel services, this documentation serves as a valuable resource to guide you through the process.\n\n### Getting Started\nTo get started, ensure you have the necessary authentication tokens and follow the endpoint descriptions to make your first API call.\n\nFor any questions or support, please refer to the contact section at the end of this documentation.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth"
      }
    },
    "schemas": {
...
      "SearchAirLegs": {
        "type": "object",
        "properties": {
          "departureDate": {
            "type": "string",
            "description": "Departure date for the flight",
            "example": "2025-06-15",
            "format": "date"
          },
          "origin": {
            "type": "string",
            "description": "Origin airport code",
            "example": "NYC"
          },
          "destination": {
            "type": "string",
            "description": "Destination airport code",
            "example": "LAX"
          }
        },
        "required": [
          "departureDate",
          "origin",
          "destination"
        ]
      },
...
      "OrderDetailDto": {
        "type": "object",
        "properties": {
          "providerCode": {
            "type": "string",
            "description": "Provider code",
            "example": "PK"
          },
          "bookingDetails": {
            "description": "bookingDetails will be here.",
            "example": [
              {
                "orderStatus": 1,
                "pnr": "abcdf",
                "orderNo": "12345",
                "firstName": "test",
                "lastName": "user"
              }
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "searchReqId": {
            "type": "string",
            "description": "Search request ID",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          },
          "searchAirLegs": {
            "description": "Search air legs information",
            "example": [
              {
                "origin": "NYC",
                "destination": "LAX"
              },
              {
                "origin": "LAX",
                "destination": "NYC"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchAirLegs"
            }
          },
          "mode": {
            "type": "string",
            "description": "Provider mode",
            "example": "Test",
            "enum": [
              "Production",
              "Test",
              "Sandbox"
            ]
          }
        },
        "required": [
          "providerCode",
          "bookingDetails",
          "searchReqId",
          "searchAirLegs",
          "mode"
        ]
      },
...
      "Passenger": {
        "type": "object",
        "properties": {
          "passengerType": {
            "type": "string",
            "description": "Type of passenger",
            "example": "ADT",
            "enum": [
              "ADT",
              "CHD",
              "INF"
            ]
          },
          "gender": {
            "type": "string",
            "description": "Passenger gender",
            "example": "M",
            "enum": [
              "M",
              "F"
            ]
          },
          "passengerDetail": {
            "description": "Passenger personal details",
            "example": {
              "title": "Mr",
              "firstName": "John",
              "middleName": "Doe",
              "lastName": "Doe"
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/PassengerDetail"
              }
            ]
          },
          "dateOfBirth": {
            "type": "string",
            "description": "Passenger date of birth",
            "example": "1990-01-15",
            "format": "date"
          },
          "document": {
            "description": "Travel document information",
            "example": {
              "documentType": "Passport",
              "documentNumber": "A1234567",
              "expiryDate": "2030-12-31",
              "country": "US"
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/Document"
              }
            ]
          },
          "nationality": {
            "type": "string",
            "description": "Passenger nationality",
            "example": "US"
          },
          "mobile": {
            "type": "string",
            "description": "Passenger mobile number",
            "example": "1234567890"
          },
          "mobileCountryCode": {
            "type": "string",
            "description": "Mobile country code",
            "example": "+1"
          }
        },
        "required": [
          "passengerType",
          "gender",
          "passengerDetail",
          "dateOfBirth",
          "nationality",
          "mobile",
          "mobileCountryCode"
        ]
      },
...
      "BookDto": {
        "type": "object",
        "properties": {
          "solutionId": {
            "type": "string",
            "description": "Solution ID for booking",
            "example": "7e3294817ad26fe6a885ca3b57312dbd"
          },
          "searchReqId": {
            "type": "string",
            "description": "Search request ID",
            "example": "efc080e1-3989-427d-abd4-edc7c04d27ab",
            "format": "uuid"
          },
          "trackingId": {
            "type": "string",
            "description": "Tracking ID",
            "example": "bdd3ff73-5ee7-49fc-bafa-34ba0666bfb7"
          },
          "providerCode": {
            "type": "string",
            "description": "Provider code",
            "example": "PK"
          },
          "fareType": {
            "type": "string",
            "description": "Fare type",
            "example": "PUBLIC",
            "enum": [
              "PUBLIC",
              "PRIVATE",
              "CORPORATE"
            ]
          },
          "airTripType": {
            "type": "string",
            "description": "airTripType needed here.",
            "example": "oneway"
          },
          "airlineType": {
            "type": "string",
            "description": "airlineType received in search result, could be LCC, Non-LCC GDS etc.",
            "example": "LCC"
          },
          "passengers": {
            "description": "Passenger information",
            "example": [
              {
                "passengerType": "ADT",
                "gender": "M",
                "passengerDetail": {
                  "title": "Mr",
                  "firstName": "John",
                  "lastName": "Doe"
                },
                "dateOfBirth": "1990-01-15",
                "document": {
                  "documentType": "P",
                  "documentNumber": "A1234567",
                  "expiryDate": "2030-12-31",
                  "country": "US"
                },
                "nationality": "US",
                "mobile": "1234567890",
                "mobileCountryCode": "+1"
              },
              {
                "passengerType": "ADT",
                "gender": "F",
                "passengerDetail": {
                  "title": "Ms",
                  "firstName": "Jane",
                  "lastName": "Doe"
                },
                "dateOfBirth": "1992-03-20",
                "document": {
                  "documentType": "P",
                  "documentNumber": "B7654321",
                  "expiryDate": "2030-12-31",
                  "country": "US"
                },
                "nationality": "US",
                "mobile": "1234567891",
                "mobileCountryCode": "+1"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Passenger"
            }
          },
          "contact": {
            "description": "Contact information",
            "example": {
              "title": "Mr",
              "firstName": "John",
              "lastName": "Doe",
              "gender": "M",
              "email": "john.doe@example.com",
              "mobile": "1234567890",
              "mobileCountryCode": "+1",
              "postalCode": "10001"
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/ContactInfo"
              }
            ]
          },
          "routes": {
            "description": "Route information - array of flight segments for each leg of the journey",
            "example": [
              [
                {
                  "airline": "AA",
                  "departureCode": "LGA",
                  "departureDate": "2026-01-15",
                  "departureTime": "09:29 AM",
                  "arrivalCode": "ORD",
                  "arrivalDate": "2026-01-15",
                  "arrivalTime": "11:10 AM",
                  "bookingCode": "Q",
                  "flightNum": "3234"
                },
                {
                  "airline": "AA",
                  "departureCode": "ORD",
                  "departureDate": "2026-01-15",
                  "departureTime": "02:14 PM",
                  "arrivalCode": "LAX",
                  "arrivalDate": "2026-01-15",
                  "arrivalTime": "04:52 PM",
                  "bookingCode": "Q",
                  "flightNum": "1204"
                }
              ]
            ],
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "gst": {
            "description": "GST details (optional - may be required based on flight/supplier)",
            "example": {
              "gstCompanyAddress": "123 Main Street",
              "gstCompanyContactNumber": "1234567890",
              "gstCompanyName": "ABC Company Ltd.",
              "gstNumber": "27AAAAA0000A1Z5",
              "gstCompanyEmail": "company@example.com"
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/GSTDetails"
              }
            ]
          }
        },
        "required": [
          "solutionId",
          "searchReqId",
          "trackingId",
          "providerCode",
          "fareType",
          "airTripType",
          "airlineType",
          "passengers",
          "contact",
          "routes"
        ]
      },
      "BookConfirmationDto": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string",
            "description": "Booking ID",
            "example": "1234567890"
          },
          "bookingLogId": {
            "type": "string",
            "description": "Booking log ID",
            "example": "1234567890"
          }
        },
        "required": [
          "bookingId",
          "bookingLogId"
        ]
      },
...
      "RevalidateDto": {
        "type": "object",
        "properties": {
          "solutionId": {
            "type": "string",
            "description": "Solution ID to revalidate",
            "example": "r5TawedGI+0dM3carFbMJrd0HGyLKfhDxLXg4GJedq2AV9JFLKQegvdbLSnClFO2Mkx7b5JAgrZdSWMarjX8lKqrnAqDCV5a3yMWYjrHknjx5zPImWcAfD9DI81sDgJjMTuz1HJ1yvECFaWGfmsXHlX4+D4kOpqO4H4yB8c2ZFKa/7JlNub+07SJ44LWbWqYeuf0k3Qz7od788m53mnZtynA4x3fcDOVMYrmn7QlH6fVwEW0HZM+htwivXaVQSoRYAP0yLTjH3DaCifTS4ZTvCSw8mmeMBN2DKEYlMwaKNRAdvuIe3x/lSMX4CYWOMRkaDNbsm5eFPiLphWK10SQUqBK+AIX8C6PfKifMNutFKaw/bmDid1/Sgxz+jTYDvFCnKg/aQpnz8vUXoCQVSynQw=="
          },
          "searchReqId": {
            "type": "string",
            "description": "Search request ID",
            "example": "efc080e1-3989-427d-abd4-edc7c04d27ab",
            "format": "uuid"
          },
          "trackingId": {
            "type": "string",
            "description": "trackingId received in search result.",
            "example": "bdd3ff73-5ee7-49fc-bafa-34ba0666bfb7"
          },
          "providerCode": {
            "type": "string",
            "description": "Provider code",
            "example": "PK"
          },
          "airTripType": {
            "type": "string",
            "description": "Type of air trip",
            "example": "OneWay",
            "enum": [
              "OneWay",
              "RoundTrip",
              "MultiCity"
            ]
          },
          "cabinClass": {
            "type": "string",
            "description": "Cabin class",
            "example": "Economy",
            "enum": [
              "Economy",
              "Business",
              "First"
            ]
          },
          "isMultiReValid": {
            "type": "boolean",
            "description": "Whether this is a multi-revalidation request",
            "example": false,
            "default": false
          },
          "groupHash": {
            "description": "Group hash information",
            "example": [
              {
                "provider": "PK",
                "netAmount": 370.7,
                "solutionId": "r5TawedGI+0dM3carFbMJrd0HGyLKfhDxLXg4GJedq2AV9JFLKQegvdbLSnClFO2Mkx7b5JAgrZdSWMarjX8lKqrnAqDCV5a3yMWYjrHknjx5zPImWcAfD9DI81sDgJjMTuz1HJ1yvECFaWGfmsXHlX4+D4kOpqO4H4yB8c2ZFKa/7JlNub+07SJ44LWbWqYeuf0k3Qz7od788m53mnZtynA4x3fcDOVMYrmn7QlH6fVwEW0HZM+htwivXaVQSoRYAP0yLTjH3DaCifTS4ZTvCSw8mmeMBN2DKEYlMwaKNRAdvuIe3x/lSMX4CYWOMRkaDNbsm5eFPiLphWK10SQUqBK+AIX8C6PfKifMNutFKaw/bmDid1/Sgxz+jTYDvFCnKg/aQpnz8vUXoCQVSynQw=="
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupHash"
            }
          },
          "paxes": {
            "description": "Passenger information",
            "example": [
              {
                "adult": 2,
                "children": 0,
                "infant": 0
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaxesInfo"
            }
          },
          "routes": {
            "description": "Route information - array of flight segments for each leg of the journey",
            "example": [
              [
                {
                  "airline": "AA",
                  "departureCode": "LGA",
                  "departureDate": "2026-06-15",
                  "departureTime": "09:29 AM",
                  "arrivalCode": "ORD",
                  "arrivalDate": "2026-06-15",
                  "arrivalTime": "11:10 AM",
                  "flightNum": "3234",
                  "bookingCode": "Q"
                },
                {
                  "airline": "AA",
                  "departureCode": "ORD",
                  "departureDate": "2026-06-15",
                  "departureTime": "02:14 PM",
                  "arrivalCode": "LAX",
                  "arrivalDate": "2026-06-15",
                  "arrivalTime": "04:52 PM",
                  "flightNum": "1204",
                  "bookingCode": "Q"
                }
              ]
            ],
            "items": {
              "type": "array"
            },
            "type": "array"
          }
        },
        "required": [
          "solutionId",
          "searchReqId",
          "trackingId",
          "providerCode",
          "airTripType",
          "cabinClass",
          "paxes",
          "routes"
        ]
      },
      "BalanceCheckDto": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "description": "Provider mode",
            "example": "Production",
            "enum": [
              "Production",
              "Test",
              "Sandbox"
            ]
          },
          "providerCode": {
            "type": "string",
            "description": "Provider code",
            "example": "PK"
          }
        },
        "required": [
          "mode",
          "providerCode"
        ]
      },
...
    }
  }
}